refactor: rename DemoMode to SandboxMode
