+++ config/releases.exs
@@ -37,4 +37,4 @@ config :radiator, :auth,
   email_from_name: System.get_env("EMAIL_FROM_NAME", "Radiator-Instance"),
   email_from_email: System.get_env("EMAIL_FROM_ADDRESS", "do_not_reply@radiator.local")
 
-config :radiator, :demo_mode, enabled: System.get_env("DEMO_MODE_ENABLED", false)
+config :radiator, :sandbox_mode, enabled: System.get_env("SANDBOX_MODE_ENABLED", false)
